Strip versions from meta.yaml. Conda build is working
